Refactor Playlist components
`src/App.tsx`
- Moved PlaylistList logic to PlaylistList.tsx

`src/components/PlaylistList.tsx`
- Created component for handling fetching of Spotify Playlists via useEffect() and returning of <PlaylistItem> components

Renamed `src/components/SpotifyPlaylist.tsx` to `src/components/PlaylistItem.tsx`
