+++ src/App.tsx
@@ -1,40 +1,16 @@
-import React, {useEffect, useState} from 'react';
 import {StyleSheet, Text} from 'react-native';
 import SpotifyAuthButton from './components/SpotifyAuthButton';
 import {useAuth} from './contexts/AuthContext';
-import {SpotifyPlaylist} from './types/SpotifyPlaylist';
-import {getPlaylists} from './services/playlistService';
-import Playlist from './components/SpotifyPlaylist';
+import PlaylistList from './components/PlaylistList';
 
 function App(): React.JSX.Element {
-  const {auth, isAuthenticated} = useAuth();
-  const [playlists, setPlaylists] = useState<SpotifyPlaylist[] | null>(null);
-
-  useEffect(() => {
-    const fetchPlaylists = async () => {
-      if (!isAuthenticated) {
-        return;
-      }
-      const data = await getPlaylists(auth!.accessToken);
-      setPlaylists(data);
-    };
-    if (isAuthenticated) {
-      fetchPlaylists();
-    } else {
-      setPlaylists(null);
-    }
-  }, [auth, isAuthenticated]);
+  const isAuthenticated = useAuth();
 
   return (
     <>
       <Text style={styles.title}>SuperSetList</Text>
       <SpotifyAuthButton />
-      {playlists &&
-        playlists.map((playlist, playlistIdx) => {
-          return (
-            <Playlist key={playlistIdx} spotifyPlaylist={playlist}></Playlist>
-          );
-        })}
+      {isAuthenticated && <PlaylistList />}
     </>
   );
 }
